moved child block list and added tests
