+++ src/app/child-dev-project/attendance/dashboard-widgets/attendance-average-dashboard/attendance-average-dashboard.component.spec.ts
@@ -1,14 +1,13 @@
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
-
 import { AttendanceAverageDashboardComponent } from "./attendance-average-dashboard.component";
 import { MatCardModule } from "@angular/material/card";
 import { MatIconModule } from "@angular/material/icon";
 import { ChildrenService } from "../../../children/children.service";
-import { ChildBlockComponent } from "../../../children/child-block/child-block.component";
 import { RouterTestingModule } from "@angular/router/testing";
 import { SchoolBlockComponent } from "../../../schools/school-block/school-block.component";
 import { of } from "rxjs";
 import { Child } from "../../../children/model/child";
+import { ChildBlockComponent } from "../../../children/child-block-list/child-block/child-block.component";
 
 describe("AttendanceAverageDashboardComponent", () => {
   let component: AttendanceAverageDashboardComponent;
