+++ src/app/child-dev-project/attendance/dashboard-widgets/attendance-warnings-dashboard/attendance-warnings-dashboard.component.spec.ts
@@ -3,12 +3,12 @@ import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 import { AttendanceWarningsDashboardComponent } from "./attendance-warnings-dashboard.component";
 import { MatCardModule } from "@angular/material/card";
 import { MatIconModule } from "@angular/material/icon";
-import { ChildBlockComponent } from "../../../children/child-block/child-block.component";
 import { RouterTestingModule } from "@angular/router/testing";
 import { ChildrenService } from "../../../children/children.service";
 import { SchoolBlockComponent } from "../../../schools/school-block/school-block.component";
 import { of } from "rxjs";
 import { Child } from "../../../children/model/child";
+import { ChildBlockComponent } from "../../../children/child-block-list/child-block/child-block.component";
 
 describe("AttendanceWarningsDashboardComponent", () => {
   let component: AttendanceWarningsDashboardComponent;
