+++ src/app/child-dev-project/attendance/dashboard-widgets/attendance-week-dashboard/attendance-week-dashboard.component.spec.ts
@@ -4,7 +4,6 @@ import { AttendanceWeekDashboardComponent } from "./attendance-week-dashboard.co
 import { MatCardModule } from "@angular/material/card";
 import { MatIconModule } from "@angular/material/icon";
 import { MatTooltipModule } from "@angular/material/tooltip";
-import { ChildBlockComponent } from "../../../children/child-block/child-block.component";
 import { AttendanceDayBlockComponent } from "../../attendance-days/attendance-day-block.component";
 import { SchoolBlockComponent } from "../../../schools/school-block/school-block.component";
 import { ChildrenService } from "../../../children/children.service";
@@ -13,6 +12,7 @@ import { Database } from "../../../../core/database/database";
 import { MockDatabase } from "../../../../core/database/mock-database";
 import { RouterTestingModule } from "@angular/router/testing";
 import { ChildPhotoService } from "../../../children/child-photo-service/child-photo.service";
+import { ChildBlockComponent } from "../../../children/child-block-list/child-block/child-block.component";
 
 describe("AttendanceWeekDashboardComponent", () => {
   let component: AttendanceWeekDashboardComponent;
