+++ src/app/child-dev-project/children/child-block-list/child-block-list.component.ts
@@ -1,8 +1,8 @@
 import { Component } from "@angular/core";
-import { OnInitDynamicComponent } from "../../../../core/view/dynamic-components/on-init-dynamic-component.interface";
+import { OnInitDynamicComponent } from "../../../core/view/dynamic-components/on-init-dynamic-component.interface";
 
 @Component({
-  selector: "app-children-list",
+  selector: "app-child-block-list",
   template: `
     <app-child-block
       *ngFor="let child of children"
@@ -10,7 +10,7 @@ import { OnInitDynamicComponent } from "../../../../core/view/dynamic-components
     ></app-child-block>
   `,
 })
-export class ChildrenListComponent implements OnInitDynamicComponent {
+export class ChildBlockListComponent implements OnInitDynamicComponent {
   public children: string[] = [];
 
   constructor() {}
