+++ src/app/child-dev-project/children/child-select/child-select.component.spec.ts
@@ -1,7 +1,6 @@
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 
 import { ChildSelectComponent } from "./child-select.component";
-import { ChildBlockComponent } from "../child-block/child-block.component";
 import { MatAutocompleteModule } from "@angular/material/autocomplete";
 import { MatFormFieldModule } from "@angular/material/form-field";
 import { MatIconModule } from "@angular/material/icon";
@@ -14,6 +13,7 @@ import { SchoolBlockComponent } from "../../schools/school-block/school-block.co
 import { of } from "rxjs";
 import { Child } from "../model/child";
 import { RouterTestingModule } from "@angular/router/testing";
+import { ChildBlockComponent } from "../child-block-list/child-block/child-block.component";
 
 function createTestData() {
   const c1 = new Child("1");
