+++ src/app/child-dev-project/children/children.module.ts
@@ -41,7 +41,6 @@ import { ChildrenService } from "./children.service";
 import { AttendanceBlockComponent } from "../attendance/attendance-block/attendance-block.component";
 import { ChildAttendanceComponent } from "../attendance/child-attendance/child-attendance.component";
 import { EntitySubrecordModule } from "../../core/entity-subrecord/entity-subrecord.module";
-import { ChildBlockComponent } from "./child-block/child-block.component";
 import { ChildrenCountDashboardComponent } from "./children-count-dashboard/children-count-dashboard.component";
 import { FormsModule, ReactiveFormsModule } from "@angular/forms";
 import { RouterModule } from "@angular/router";
@@ -84,6 +83,8 @@ import { FormComponent } from "./child-details/form/form.component";
 import { ListAttendanceComponent } from "./children-list/list-attendance/list-attendance.component";
 import { EntityListModule } from "../../core/entity-list/entity-list.module";
 import { SchoolBlockWrapperComponent } from "./children-list/school-block-wrapper/school-block-wrapper.component";
+import { ChildBlockListComponent } from "./child-block-list/child-block-list.component";
+import { ChildBlockComponent } from "./child-block-list/child-block/child-block.component";
 
 @NgModule({
   imports: [
@@ -156,6 +157,7 @@ import { SchoolBlockWrapperComponent } from "./children-list/school-block-wrappe
     FormComponent,
     ListAttendanceComponent,
     SchoolBlockWrapperComponent,
+    ChildBlockListComponent,
   ],
   providers: [
     ChildrenService,
