--- src/app/child-dev-project/notes/notes-manager/children-list/children-list.component.spec.ts
@@ -1,24 +1,0 @@
-import { async, ComponentFixture, TestBed } from "@angular/core/testing";
-
-import { ChildrenListComponent } from "./children-list.component";
-
-describe("ChildrenListComponent", () => {
-  let component: ChildrenListComponent;
-  let fixture: ComponentFixture<ChildrenListComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ChildrenListComponent],
-    }).compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ChildrenListComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it("should create", () => {
-    expect(component).toBeTruthy();
-  });
-});
